==> notebooks/1_fetch_data.ipynb <==
Step 1 '1_fetch_data': This notebook handles the initial data acquisition phase of the project. It automates the process of downloading a large dataset of song lyrics and genres from a Kaggle dataset using the Kaggle CLI. The notebook also includes steps for extracting the downloaded zip file and organizing the data into the appropriate directory for subsequent processing.

import os
import subprocess

# Set the desired paths
dataset_path = "nikhilnayak123/5-million-song-lyrics-dataset"
file_name = "ds2.csv"
destination_path = "../data/pristine/"

# Ensure the directory exists
os.makedirs(destination_path, exist_ok=True)

# Download the specific file using kaggle CLI
subprocess.run([
    "kaggle", "datasets", "download", "-d", dataset_path, "-f", file_name, "-p", destination_path
])


import zipfile
import os

# Define paths
zip_path = "../data/pristine/ds2.csv.zip"
destination_path = "../data/pristine/"

# Unzip the file
# This takes a long time to complete!!!
with zipfile.ZipFile(zip_path, 'r') as zip_ref:
    zip_ref.extractall(destination_path)

# Remove the zip file
os.remove(zip_path)
