==> notebooks/3_data_cleaning.ipynb <==
Step 3: '3_data_cleaning' : This step focused on cleaning the data for our topic mining. The cleaning step removes nonsensical “words” from the lyrics such as words in brackets, “\n”, and non-english words. Also, it removes all empty songs that do not contain any lyrics.

import re
import pandas as pd

file_path = '../data/2_ds2_trimmed.csv'

df = pd.read_csv(file_path)
df = df.dropna(subset=['tag', 'lyrics'])

# Used this to test different chunks of data

import pandas as pd
column_names = pd.read_csv(file_path, nrows=0).columns

chunk_size = 10**5  # Define your chunk size
chunk_number = 3     # Specify which chunk you want (e.g., 2nd chunk here)

# Calculate the number of rows to skip
rows_to_skip = chunk_size * (chunk_number - 1)+1

# Load the next chunk with skiprows
df = pd.read_csv(file_path, skiprows=rows_to_skip, nrows=chunk_size,  header=None,names=column_names)
df['lyrics']=df['lyrics'].astype(str)
# Display a sample to verify
df.head()

df.head()

def clean_lyrics(text):
    # Remove all text within square brackets
    text = re.sub(r'\[.*?\]', '', text)
    text = re.sub(r'\{.*?\}', '', text)
    text = re.sub(r'\(.*?\)', '', text)
    # Remove any remaining newline characters
    text = text.replace('\n', ' ')
    # Remove leading/trailing whitespace
    text = text.strip()
    return text

# clean out nonuseful words
df['lyrics']=df['lyrics'].apply(clean_lyrics)
# Keep only lyrics with 20 or more words
clean_df = df[df['lyrics'].str.split().str.len() >= 20]

# note gets rid of short songs, zero lyrics, intrumentals, and non lric words such as intro chorus etc

clean_df.head()

# Sample of these short lyrics to inspect
#sampled_short_lyrics = short_lyrics.sample(15, random_state=15)

# Display the results
#for index, lyric in enumerate(sampled_short_lyrics['lyrics'], 1):
#    print(f"Lyrics Sample {index}:\n{lyric}\n{'-'*50}")


from langdetect import detect, DetectorFactory
from langdetect.lang_detect_exception import LangDetectException

# Ensure consistent results from langdetect
DetectorFactory.seed = 0

def is_english(text):
    try:
        # Detect language of text
        return detect(text) == 'en'
    except LangDetectException:
        # If language detection fails, treat it as non-English
        return False


# #This was a check to see if it sorted correctly
# # Note this can take a long time with large inputs

# # Sample 100 random lyrics along with titles
# sampled_lyrics = df[['title', 'lyrics']].sample(100, random_state=5)

# # Check each sampled lyric for English language
# sampled_lyrics['is_english'] = sampled_lyrics['lyrics'].apply(is_english)

# # Filter for non-English lyrics in the sampled set
# non_english_lyrics = sampled_lyrics[sampled_lyrics['is_english'] == False]

# # Display non-English sampled lyrics
# print("Sampled Lyrics Detected as Non-English:")
# for index, row in non_english_lyrics.iterrows():
#     print(f"Title: {row['title']}\nLyrics:\n{row['lyrics']}\n{'-'*50}")


# note this can take a very large time recommend trying your computer with a sample first.

clean_df = clean_df[clean_df['lyrics'].apply(is_english)]


clean_df.to_csv("../data/3_ds3_cleaned.csv", index=False)
